refactor(activation): rename intent-resolver to button-intent-resolver
Mirror the resolveButtonIntent export and the button-activation.ts sibling;
the file↔function naming now matches the board-key-resolver/resolveBoardKey
convention used elsewhere.
